# perf: update whatwg_url package to latest commit